+++ lib/utils.ts
@@ -295,8 +295,3 @@ export const sortBytesArray = _array =>
 
 export const formatRound = (num, decimals = 2) =>
   isNaN(num) ? '----' : round(num, decimals).toLocaleString({}, { minimumFractionDigits: decimals });
-
-export const getVotingWeightCopy = (isVotingDelegate: boolean): string =>
-  isVotingDelegate
-    ? 'Your voting weight is made up of the MKR delegated to your delegate contract.'
-    : 'Your voting weight is made up of the MKR in your wallet, vote proxy, and voting contract. This amount is applied to all polls you vote on.';
